+++ docs/CHANGES.rst
@@ -8,7 +8,9 @@ Changelog
   [sgeulette]
 - Added MailingLoopTemplate type and mailing_loop_template field on ConfigurablePODTemplate.
   [sgeulette]
-- Added 'mailing-loop-persistent-document-generation' to manage mailing loop generation
+- Added 'mailing-loop-persistent-document-generation' view to manage mailing loop generation
+  [sgeulette]
+- Added helper method to manage context
   [sgeulette]
 
 2.0.8 (2017-08-02)